finished, writing UT

+++ src/server/util/file_io.cc
@@ -29,7 +29,7 @@ void PutFixed32(std::string* dst, uint32_t value) {
   dst->append(buf, sizeof(buf));
 }
 
-void PutFixed64(std::string* dst, uint32_t value) {
+void PutFixed64(std::string* dst, uint64_t value) {
   char buf[sizeof(value)];
   EncodeFixed64(buf, value);
   dst->append(buf, sizeof(buf));
